Refactor format_section_override out of style-spec (#9543)
