+++ src/style/format_section_override.js
@@ -1,14 +1,16 @@
 // @flow
 
 import assert from 'assert';
-import type {Expression} from '../expression';
-import type EvaluationContext from '../evaluation_context';
-import type {Type} from '../types';
-import type {ZoomConstantExpression} from '../../expression';
-import {NullType} from '../types';
-import {PossiblyEvaluatedPropertyValue} from '../../../style/properties';
-import {register} from '../../../util/web_worker_transfer';
+import type {Expression} from '../style-spec/expression/expression';
+import type EvaluationContext from '../style-spec/expression/evaluation_context';
+import type {Type} from '../style-spec/expression/types';
+import type {ZoomConstantExpression} from '../style-spec/expression';
+import {NullType} from '../style-spec/expression/types';
+import {PossiblyEvaluatedPropertyValue} from './properties';
+import {register} from '../util/web_worker_transfer';
 
+// This is an internal expression class. It is only used in GL JS and
+// has GL JS dependencies which can break the standalone style-spec module
 export default class FormatSectionOverride<T> implements Expression {
     type: Type;
     defaultValue: PossiblyEvaluatedPropertyValue<T>;
